fixed rubocop offenses
